edit all functional

+++ config/documents/views.py
@@ -1,6 +1,7 @@
 import csv
 from datetime import datetime
 
+from django.conf import settings
 from django.db.models import Q
 from django.http import HttpResponse
 from django.shortcuts import render, redirect
@@ -11,10 +12,44 @@ from .forms import DocumentForm
 from .models import DocumentModel
 
 
-class DocumentsView(View):
+class HomeView(View):
 
     def get(self, request):
-        documents = DocumentModel.objects.exclude(~Q(user=request.user.pk), private=True)
+        if not request.user.is_authenticated and request.user.status:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
+        return redirect('incoming')
+
+
+class DocumentsIncoming(View):
+
+    def get(self, request):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
+        documents = DocumentModel.objects.filter(user_to__pk=request.user.pk, archived=False)
+        context = {
+            'documents': documents,
+        }
+        return render(request, 'documents/index.html', context)
+
+
+class DocumentsArchive(View):
+
+    def get(self, request):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
+        documents = DocumentModel.objects.filter(user_from__pk=request.user.pk, archived=True)
+        context = {
+            'documents': documents,
+        }
+        return render(request, 'documents/archive.html', context)
+
+
+class DocumentsOutcoming(View):
+
+    def get(self, request):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
+        documents = DocumentModel.objects.filter(user_from__pk=request.user.pk, archived=False)
         context = {
             'documents': documents,
         }
@@ -22,6 +57,25 @@ class DocumentsView(View):
 
 
 class DocumetnsAdd(View):
+    '''
+    name = models.CharField('Название документа', max_length=255, null=True)
+    theme = models.CharField('Вид документа', max_length=255, null=True)
+    number = models.IntegerField('Номер договора', null=True)
+    type = models.CharField('Тип документа', max_length=255, null=True)
+    description = models.TextField('Краткое содержание документа', null=True)
+    status = models.CharField('Статус документа', choices=DOCUMENT_STATUSES, max_length=255, default=ADDED)
+
+    load_date = models.DateTimeField(verbose_name='Дата добавления', auto_now_add=True)
+    ending_date = models.DateTimeField(verbose_name='Дата окончания', null=True, default=datetime.now())
+    archived = models.BooleanField('В архиве', default=False)
+    user_from = models.ForeignKey(CustomUser,
+                                  verbose_name='От кого',
+                                  on_delete=models.CASCADE,
+                                  null=True,
+                                  related_name='user_from')
+    user_to = models.ManyToManyField(CustomUser, verbose_name='Кому', null=True, related_name='user_to')
+    file = models.FileField(verbose_name='Электронный файл докуента', upload_to='documents/', null=True)
+    '''
 
     def post(self, request):
         form = DocumentForm(request.POST, request.FILES)
@@ -30,14 +84,16 @@ class DocumetnsAdd(View):
             document = DocumentModel.objects.create(file=request.FILES['file'])
             document.name = form.cleaned_data['name']
             document.number = form.cleaned_data['number']
-            document.user = request.user
             document.type = form.cleaned_data['type']
-            document.author = form.cleaned_data['author']
-            document.doc_from = form.cleaned_data['doc_from']
             document.description = form.cleaned_data['description']
             document.theme = form.cleaned_data['theme']
+            document.user_from = request.user
+            document.user_to = form.cleaned_data['user_to']
+            document.status = DocumentModel.ADDED
+            document.action = form.cleaned_data['action']
+            document.ending_date = form.cleaned_data['ending_date']
             document.save()
-            return redirect('documents')
+            return redirect('outcoming')
 
         context = {
             'form': form
@@ -45,6 +101,8 @@ class DocumetnsAdd(View):
         return render(request, 'documents/add.html', context)
 
     def get(self, request):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
         form = DocumentForm(data=request.POST or None)
         context = {
             'form': form,
@@ -52,42 +110,56 @@ class DocumetnsAdd(View):
         return render(request, 'documents/add.html', context)
 
 
-class DocumentsChange(View):
-
-    def post(self, request, **kwargs):
-        form = DocumentForm(request.POST, request.FILES)
-        if form.is_valid():
-            form.save(commit=False)
-            document = DocumentModel.objects.get(pk=kwargs['pk'])
-            document.name = form.cleaned_data['name']
-            document.number = form.cleaned_data['number']
-            document.user = request.user
-            document.type = form.cleaned_data['type']
-            document.author = form.cleaned_data['author']
-            document.doc_from = form.cleaned_data['doc_from']
-            document.description = form.cleaned_data['description']
-            document.theme = form.cleaned_data['theme']
-            document.file = request.FILES['file']
-            document.save()
-            return redirect('documents')
-
-        context = {
-            'form': form
-        }
-        return render(request, 'documents/add.html', context)
+class DocumentsToArchive(View):
 
     def get(self, request, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
         document = DocumentModel.objects.get(pk=kwargs['pk'])
-        form = DocumentForm(instance=document)
-        context = {
-            'form': form,
-        }
-        return render(request, 'documents/add.html', context)
+        document.archived = True
+        document.save()
+        return redirect('outcoming')
+
+
+class DocumentsSubmit(View):
+
+    def get(self, request, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
+        document = DocumentModel.objects.get(pk=kwargs['pk'])
+        document.submited = True
+        document.status = DocumentModel.SUBMITED
+        document.save()
+        return redirect('incoming')
+
+
+class DocumentsTerminate(View):
+
+    def get(self, request, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
+        document = DocumentModel.objects.get(pk=kwargs['pk'])
+        document.status = DocumentModel.TERMINATED
+        document.save()
+        return redirect('incoming')
+
+
+class DocumentsRead(View):
+
+    def get(self, request, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
+        document = DocumentModel.objects.get(pk=kwargs['pk'])
+        document.status = DocumentModel.READ
+        document.save()
+        return redirect('incoming')
 
 
 class DocumentsDelete(View):
 
     def get(self, request, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
         document = DocumentModel.objects.get(pk=kwargs['pk'])
         document.delete()
 
@@ -97,7 +169,8 @@ class DocumentsDelete(View):
 class DocumentsFilter(View):
 
     def get(self, request, **kwargs):
-
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
         choice = request.GET['choice']
         if choice == 'name':
             documents = DocumentModel.objects.filter(name__icontains=request.GET['q'])
@@ -124,7 +197,8 @@ class DocumentsFilter(View):
 class DocumentsSort(View):
 
     def get(self, request, **kwargs):
-
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
         choice = request.GET['choice']
         up_down = request.GET['up_down']
 
@@ -148,10 +222,14 @@ class DescriptionView(View):
 class CreateReport(View):
 
     def get(self, request):
-
+        if not request.user.is_authenticated:
+            return redirect('%s?next=%s' % (settings.LOGIN_URL, request.path))
         date = datetime.strptime(request.GET['date'], "%Y-%m-%d")
-        documents = DocumentModel.objects.filter(change_date__year=date.year, change_date__month=date.month,
-                                                 change_date__day=date.day)
+        doc_from = DocumentModel.objects.filter(load_date__year=date.year, load_date__month=date.month,
+                                                load_date__day=date.day, user_from__pk=request.user.pk)
+        doc_to = DocumentModel.objects.filter(load_date__year=date.year, load_date__month=date.month,
+                                              load_date__day=date.day, user_to__pk=request.user.pk)
+        documents = doc_from.union(doc_to)
 
         if not documents:
             documents = DocumentModel.objects.all()
@@ -165,8 +243,12 @@ class CreateReport(View):
 
             data = []  # Окончательный список данных, которые будут возвращены
             for row in documents:
-                source_data = [row.number, row.name, row.type, row.author, row.doc_from, row.theme,
-                               f'{row.user.first_name} {row.user.last_name}', f'{row.change_date.strftime("%Y.%m.%d")}'
+                source_data = [row.number, row.name, row.type, row.theme,
+                               f'{row.user_from.first_name} {row.user_from.last_name}',
+                               f'{row.user_to.first_name} {row.user_to.last_name}',
+                               f'{"да" if row.archived else "нет"}',
+                               f'{row.load_date.strftime("%Y.%m.%d")}',
+                               f'{row.ending_date.strftime("%Y.%m.%d")}'
                                ]
                 data.append(source_data)  # Добавить каждую строку списка данных в большой список
 
@@ -178,11 +260,12 @@ class CreateReport(View):
                 'Номер документа',
                 'Название документа',
                 'Тип документа',
-                'Автор документа',
-                'Отдел откуда документ',
                 'Тема документа',
-                'ФИО добавил',
-                'Дата добавления/изменения',
+                'От кого',
+                'Кому',
+                'В архиве',
+                'Дата создания',
+                'Срок выполнения',
             ])
             for row in data:
                 writer.writerow(row)  # Cycle для записи информации базы данных
